requirements.txt
Django==3.2.6
django-mptt==0.12.0
djangorestframework==3.12.4
psycopg2-binary==2.9.1
python-dotenv==0.19.0
drf-yasg